@@ -36,7 +36,8 @@
     "@istanbuljs/nyc-config-typescript": "^1.0.2",
     "@types/chai": "^4.3.1",
     "@types/mocha": "^9.1.1",
-    "@types/node": "^17.0.34",
+    "@types/node": "^18.11.9",
+    "@types/react": "^16.9.49",
     "@types/react-dom": "^18.0.4",
     "@types/sinon": "^10.0.11",
     "@types/sinon-chai": "^3.2.8",
@@ -57,9 +58,9 @@
     "sinon": "^14.0.0",
     "sinon-chai": "^3.7.0",
     "source-map-support": "^0.5.21",
-    "ts-node": "^10.7.0",
-    "typedoc": "^0.22.15",
-    "typescript": "^4.6.4"
+    "ts-node": "^10.9.1",
+    "typedoc": "^0.23.21",
+    "typescript": "^4.9.3"
   },
   "files": [
     "dist",
