@@ -1,29 +1,18 @@
 import React from "react";
-import { Route, Link } from "react-router-dom";
-import { withRouter } from "react-router";
+import { Link } from "react-router-dom";
 
 const NavItem = props => {
-  const { to, exact, strict, children } = props;
+  const { to, children } = props;
   return (
-    <Route
-      path={to}
-      exact={exact}
-      strict={strict}
-      children={({ location, match }) => {
-        const cn = match ? "active" : null;
-        return (
-          <li className={cn}>
-            <Link to={to}>
-              <div>{children}</div>
-            </Link>
-          </li>
-        );
-      }}
-    />
+    <li>
+      <Link to={to}>
+        <div>{children}</div>
+      </Link>
+    </li>
   );
 };
 
-const Navigation = () => {
+export const Navigation = () => {
   return (
     <nav className="navbar navbar-inverse">
       <div className="container-fluid">
@@ -65,7 +54,3 @@ const Navigation = () => {
     </nav>
   );
 };
-
-const RouterNavigation = withRouter(Navigation);
-
-export { RouterNavigation as Navigation };
